requirements.txt
pytest
requests
PyYAML
allure-pytest
selenium
webdriver-manager
openpyxl
pymysql
